app icon, formatting test
