+++ app.py
@@ -1,12 +1,15 @@
-from PySide6 import QtWidgets
+from PySide6 import QtWidgets, QtGui
 import sys
 from gui.placeholder import MainWindow
 
 if __name__ == "__main__":
     app = QtWidgets.QApplication([])
 
+    app_icon = QtGui.QIcon("assets/openposter.ico")
+    app.setWindowIcon(app_icon)
+
     widget = MainWindow()
-    widget.resize(800, 600)
+    widget.resize(1600, 900)
     widget.show()
 
     sys.exit(app.exec())
